Port to Lustre 2.6, is compatible with Lustre 2.1, 2.4 and 2.5
FSAL_LUSTRE is accessing LUSTRE via the "fid path" way. Every accessed object is associated with a fid (similar to a NFS FH in some ways).
For every object, a special and immutable path can be build to access the object. This path is "<mount point>/.lustre/fids/<object's fid>".
When trying to lookup parent from <mntpoint>/.lustre/fid the standard POSIX API points to <mntpoint>/.lustre/fid, making this special
directory the parent of every object.
To deal with that, FSAL_LUSTRE embeds a very small piece of Lustre kernel code to compute the paren correctly with proper consideration
of the filestsyem trees. The related code is located in src/FSAL/FSAL_LUSTRE/lustre_extended_types.h .
From version 2.4 to 2.6, the required code to embed has change, but with backward compatibilty.
This commits updates the embedded piece of code and makes it compatible with very version of Lustre 2.x from 2.1 to 2.6 .

Change-Id: I8680d27cbf155f6e0c5b3115fb1b276742fb4aed
Signed-off-by: Philippe DENIEL <philippe.deniel@cea.fr>