cmake: use the environment DESTDIR override for python installs
In several packaging steps, the DESTDIR is specified to install
the results to a local directory without changing the prefix
of the compilation.

So, in this case use the DESTDIR for the python installation
command to avoid any attempts to install to the root "/usr"
path when doing package installs.

Signed-off-by: Charles Hardin <ckhardin@gmail.com>