cmake: allow the FSAL_DESTINATION to be overwritten
In certain environments, the library pathes are no longer
lib, lib32, and lib64 - such as Debian based systems that support
multiarch where the library path for the 64-bit architectures
are actually /usr/lib/x86_64-linux-gnu/ganesha and use the gnu
tuple based definition for the library.

So, to maintain backwards compatibility just allow an
override to cmake that can install the FSAL modules into
a different location.

Usage:

$ cmake -DFSAL_DESTINATION=lib/x86_64-linux-gnu/ganesha

Signed-off-by: Charles Hardin <ckhardin@gmail.com>