cmake: add a debian cmake build configuration
In the long run, the cmake packager can be used to create the debian
package by adding all of the appropriate glue definitions into this
file.

In the short term, this configuration is just referenced from
the debian/rules to trampoline from a "git-buildpackage" or a
"dpkg-buildpackage" command into the debian specific build
configuration.

Signed-off-by: Charles Hardin <ckhardin@gmail.com>